+++ src/main/java/novamachina/novacore/NovaCore.java
@@ -2,7 +2,7 @@ package novamachina.novacore;
 
 import static novamachina.novacore.NovaCore.MOD_ID;
 
-import net.minecraftforge.fml.common.Mod;
+import net.neoforged.fml.common.Mod;
 
 @Mod(MOD_ID)
 public class NovaCore {