+++ src/main/java/novamachina/novacore/bootstrap/NeoforgeFluidRegistry.java
@@ -1,0 +1,15 @@
+package novamachina.novacore.bootstrap;
+
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.BuiltInRegistries;
+import novamachina.novacore.core.IRegistry;
+import novamachina.novacore.world.level.material.FluidDefinition;
+
+public class NeoforgeFluidRegistry implements IRegistry<FluidDefinition<?, ?, ?>> {
+  public NeoforgeFluidRegistry() {}
+
+  public void register(FluidDefinition<?, ?, ?> entry) {
+    Registry.register(BuiltInRegistries.FLUID, entry.getId(), entry.getStillFluid());
+    Registry.register(BuiltInRegistries.FLUID, entry.getFlowingId(), entry.getFlowingFluid());
+  }
+}