+++ src/main/java/novamachina/novacore/bootstrap/NeoforgeLootModifierRegistry.java
@@ -1,0 +1,17 @@
+package novamachina.novacore.bootstrap;
+
+import net.minecraft.core.Registry;
+import net.neoforged.neoforge.common.loot.LootModifier;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+import novamachina.novacore.common.loot.modifier.LootModifierDefinition;
+import novamachina.novacore.core.IRegistry;
+
+public class NeoforgeLootModifierRegistry
+    implements IRegistry<LootModifierDefinition<? extends LootModifier>> {
+  public NeoforgeLootModifierRegistry() {}
+
+  public void register(LootModifierDefinition<? extends LootModifier> entry) {
+    Registry.register(
+        NeoForgeRegistries.GLOBAL_LOOT_MODIFIER_SERIALIZERS, entry.getId(), entry.getCodec());
+  }
+}