+++ src/main/java/novamachina/novacore/data/recipes/ISubRecipeProvider.java
@@ -1,8 +1,7 @@
 package novamachina.novacore.data.recipes;
 
-import java.util.function.Consumer;
-import net.minecraft.data.recipes.FinishedRecipe;
+import net.minecraft.data.recipes.RecipeOutput;
 
 public interface ISubRecipeProvider {
-  void addRecipes(Consumer<FinishedRecipe> consumer);
+  void addRecipes(RecipeOutput consumer);
 }