+++ src/main/java/novamachina/novacore/util/FluidStackUtils.java
@@ -1,31 +1,28 @@
 package novamachina.novacore.util;
 
 import com.google.gson.JsonElement;
-import com.google.gson.JsonNull;
 import com.google.gson.JsonObject;
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
+import com.mojang.serialization.JsonOps;
+import java.util.Optional;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.GsonHelper;
 import net.minecraft.world.level.material.Fluid;
-import net.minecraftforge.common.util.JsonUtils;
-import net.minecraftforge.fluids.FluidStack;
-import net.minecraftforge.fluids.FluidType;
-import net.minecraftforge.registries.ForgeRegistries;
+import net.neoforged.neoforge.common.util.JsonUtils;
+import net.neoforged.neoforge.fluids.FluidStack;
+import net.neoforged.neoforge.fluids.FluidType;
+import org.slf4j.Logger;
 
 public class FluidStackUtils {
 
+  private static final Logger log = org.slf4j.LoggerFactory.getLogger(FluidStackUtils.class);
+
   private FluidStackUtils() {}
 
-  @Nullable
-  public static FluidStack deserialize(@Nonnull final JsonObject jsonObject) {
-    @Nullable
-    final Fluid fluid =
-        ForgeRegistries.FLUIDS.getValue(
+  public static FluidStack deserialize(JsonObject jsonObject) {
+    Fluid fluid =
+        BuiltInRegistries.FLUID.get(
             new ResourceLocation(GsonHelper.getAsString(jsonObject, "fluid")));
-    if (fluid == null) {
-      return null;
-    }
     FluidStack fluidStack = new FluidStack(fluid, FluidType.BUCKET_VOLUME);
     if (GsonHelper.isValidNode(jsonObject, "tag")) {
       fluidStack.setTag(JsonUtils.readNBT(jsonObject, "tag"));
@@ -34,20 +31,9 @@ public class FluidStackUtils {
     return fluidStack;
   }
 
-  @Nonnull
-  public static JsonElement serialize(@Nullable final FluidStack fluidStack) {
-    if (fluidStack == null) {
-      return JsonNull.INSTANCE;
-    }
-    JsonObject jsonObject = new JsonObject();
-    if (ForgeRegistries.FLUIDS.getKey(fluidStack.getFluid()) != null) {
-      jsonObject.addProperty(
-          "fluid", ForgeRegistries.FLUIDS.getKey(fluidStack.getFluid()).toString());
-    }
-    if (fluidStack.hasTag()) {
-      jsonObject.addProperty("tag", fluidStack.getTag().toString());
-    }
-    jsonObject.addProperty("amount", fluidStack.getAmount());
-    return jsonObject;
+  public static Optional<JsonElement> serialize(FluidStack fluidStack) {
+    return FluidStack.CODEC
+        .encodeStart(JsonOps.INSTANCE, fluidStack)
+        .resultOrPartial(error -> log.error("Unable to encode FluidStack"));
   }
 }