+++ src/main/java/novamachina/novacore/util/TankUtil.java
@@ -1,6 +1,5 @@
 package novamachina.novacore.util;
 
-import javax.annotation.Nonnull;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
@@ -8,41 +7,33 @@ import net.minecraft.world.item.alchemy.PotionUtils;
 import net.minecraft.world.item.alchemy.Potions;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.material.Fluids;
-import net.minecraftforge.fluids.FluidStack;
-import net.minecraftforge.fluids.FluidType;
-import net.minecraftforge.fluids.capability.IFluidHandler;
+import net.neoforged.neoforge.fluids.FluidStack;
+import net.neoforged.neoforge.fluids.capability.IFluidHandler;
+import net.neoforged.neoforge.fluids.capability.IFluidHandler.FluidAction;
 
 public class TankUtil {
-
   private static final ItemStack WATER_BOTTLE;
-
-  static {
-    WATER_BOTTLE = PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.WATER);
-  }
 
   private TankUtil() {}
 
   public static boolean drainWaterFromBottle(
-      @Nonnull final BlockEntity tileEntity,
-      @Nonnull final Player player,
-      @Nonnull final IFluidHandler tank) {
+      BlockEntity tileEntity, Player player, IFluidHandler tank) {
     int waterAmount = 333;
-    if (tank.getFluidInTank(0).getAmount() % FluidType.BUCKET_VOLUME == 666) {
+    if (tank.getFluidInTank(0).getAmount() % 1000 == 666) {
       waterAmount = 334;
     }
+
     if (player.getMainHandItem().getItem() == Items.POTION
         && WATER_BOTTLE.getTag() != null
         && WATER_BOTTLE.getTag().equals(player.getMainHandItem().getTag())) {
       FluidStack water = new FluidStack(Fluids.WATER, waterAmount);
-
-      if (tank.fill(water, IFluidHandler.FluidAction.SIMULATE) == water.getAmount()
+      if (tank.fill(water, FluidAction.SIMULATE) == water.getAmount()
           && player.addItem(new ItemStack(Items.GLASS_BOTTLE))) {
-
         if (!player.isCreative()) {
           player.getMainHandItem().shrink(1);
         }
-        tank.fill(water, IFluidHandler.FluidAction.EXECUTE);
 
+        tank.fill(water, FluidAction.EXECUTE);
         tileEntity.setChanged();
         return true;
       }
@@ -51,15 +42,13 @@ public class TankUtil {
     return false;
   }
 
-  // TODO: Figure out how to completely fill and drain
   public static boolean drainWaterIntoBottle(
-      @Nonnull final BlockEntity tileEntity,
-      @Nonnull final Player player,
-      @Nonnull final IFluidHandler tank) {
+      BlockEntity tileEntity, Player player, IFluidHandler tank) {
     int waterAmount = 333;
-    if (tank.getFluidInTank(0).getAmount() % FluidType.BUCKET_VOLUME == 334) {
+    if (tank.getFluidInTank(0).getAmount() % 1000 == 334) {
       waterAmount = 334;
     }
+
     if (player.getMainHandItem().getItem() == Items.GLASS_BOTTLE
         && tank.getFluidInTank(0).getFluid() != null
         && tank.getFluidInTank(0).getAmount() >= waterAmount
@@ -68,10 +57,16 @@ public class TankUtil {
       if (!player.isCreative()) {
         player.getMainHandItem().shrink(1);
       }
-      tank.drain(waterAmount, IFluidHandler.FluidAction.EXECUTE);
+
+      tank.drain(waterAmount, FluidAction.EXECUTE);
       tileEntity.setChanged();
       return true;
+    } else {
+      return false;
     }
-    return false;
+  }
+
+  static {
+    WATER_BOTTLE = PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.WATER);
   }
 }