+++ src/main/java/novamachina/novacore/world/item/crafting/Recipe.java
@@ -1,36 +1,22 @@
 package novamachina.novacore.world.item.crafting;
 
-import com.google.common.base.Preconditions;
-import lombok.Getter;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import org.checkerframework.checker.nullness.qual.NonNull;
 
-@Getter
 public abstract class Recipe implements net.minecraft.world.item.crafting.Recipe<Container> {
-
-  private final ResourceLocation id;
-
-  protected Recipe(ResourceLocation id) {
-    Preconditions.checkNotNull(id, "Id cannot be null");
-    this.id = id;
-  }
 
   public abstract void write(FriendlyByteBuf buffer);
 
   @Override
-  public boolean matches(@NonNull Container pContainer, @NonNull Level pLevel) {
+  public boolean matches(Container pContainer, Level pLevel) {
     return false;
   }
 
   @Override
-  @NonNull
-  public ItemStack assemble(
-      @NonNull Container pContainer, @NonNull RegistryAccess pRegistryAccess) {
+  public ItemStack assemble(Container pContainer, RegistryAccess pRegistryAccess) {
     return ItemStack.EMPTY;
   }
 
@@ -40,8 +26,7 @@ public abstract class Recipe implements net.minecraft.world.item.crafting.Recipe
   }
 
   @Override
-  @NonNull
-  public ItemStack getResultItem(@NonNull RegistryAccess pRegistryAccess) {
+  public ItemStack getResultItem(RegistryAccess pRegistryAccess) {
     return ItemStack.EMPTY;
   }
 }